+++ libs/cardano-ledger-api/testlib/Test/Cardano/Ledger/Api/Upgrade.hs
@@ -17,6 +17,7 @@ import Cardano.Ledger.Core
 import Cardano.Ledger.MemoBytes (EqRaw (eqRaw))
 import Data.Default (Default (def))
 import qualified Prettyprinter as Pretty
+import Test.Cardano.Ledger.Api.Arbitrary ()
 import Test.Cardano.Ledger.Binary.RoundTrip
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Arbitrary ()